Moving 4 body
