+++ include/goofit/PDFs/physics/Amp4BodyBase.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <goofit/PDFs/physics/AmpNBodyBase.h>
+
+namespace GooFit {
+
+class Amp4BodyBase : public AmpNBodyBase {
+  public:
+    using AmpNBodyBase::AmpNBodyBase;
+};
+
+} // namespace GooFit
